+++ tests/core/tests/admin_integration/test_action_export.py
@@ -1,3 +1,4 @@
+import warnings
 from datetime import datetime
 from unittest import mock
 from unittest.mock import MagicMock, PropertyMock, patch
@@ -103,7 +104,9 @@ class ExportActionAdminIntegrationTest(AdminTestMixin, TestCase):
             **self.resource_fields_payload,
         }
         date_str = datetime.now().strftime("%Y-%m-%d")
-        response = self.client.post(self.category_export_url, data)
+        with warnings.catch_warnings():
+            warnings.filterwarnings("ignore", category=DeprecationWarning)
+            response = self.client.post(self.category_export_url, data)
         self.assertEqual(response.status_code, 200)
         self.assertTrue(response.has_header("Content-Disposition"))
         self.assertEqual(response["Content-Type"], "text/csv")
@@ -148,6 +151,79 @@ class ExportActionAdminIntegrationTest(AdminTestMixin, TestCase):
                 f"{self.cat1.id} is not one of the available choices.",
                 response.content.decode(),
             )
+
+    def test_export_admin_action_with_restricted_pks_deprecated(self):
+        data = {
+            "format": "0",
+            "export_items": [str(self.cat1.id)],
+            **self.resource_fields_payload,
+        }
+        with self.assertWarnsRegex(
+            DeprecationWarning,
+            r"The 'get_valid_export_item_pks\(\)' method in "
+            "core.admin.CategoryAdmin is deprecated and will be removed "
+            "in a future release",
+        ):
+            self.client.post(self.category_export_url, data)
+
+    def _perform_export_action_calls_modeladmin_get_queryset_test(self, data):
+        # Issue #1864
+        # ModelAdmin's get_queryset should be used in the ModelAdmin mixins
+        with mock.patch(
+            "core.admin.CategoryAdmin.get_queryset"
+        ) as mock_modeladmin_get_queryset, mock.patch(
+            "import_export.admin.ExportMixin.get_data_for_export"
+        ) as mock_get_data_for_export:
+            mock_queryset = mock.MagicMock(name="MockQuerySet")
+            mock_queryset.filter.return_value = mock_queryset
+            mock_queryset.order_by.return_value = mock_queryset
+
+            mock_modeladmin_get_queryset.return_value = mock_queryset
+
+            self.client.post(self.category_export_url, data)
+
+            mock_modeladmin_get_queryset.assert_called()
+            mock_get_data_for_export.assert_called()
+
+            args, kwargs = mock_get_data_for_export.call_args
+            mock_get_data_for_export.assert_called_with(
+                args[0], mock_queryset, **kwargs
+            )
+
+    def test_export_action_calls_modeladmin_get_queryset(self):
+        # Issue #1864
+        # Test with specific export items
+
+        data = {
+            "format": "0",
+            "export_items": [str(self.cat1.id)],
+            **self.resource_fields_payload,
+        }
+        with warnings.catch_warnings():
+            warnings.filterwarnings("ignore", category=DeprecationWarning)
+            self._perform_export_action_calls_modeladmin_get_queryset_test(data)
+
+    def test_export_action_calls_modeladmin_get_queryset_all_items(self):
+        # Issue #1864
+        # Test without specific export items
+
+        data = {
+            "format": "0",
+            **self.resource_fields_payload,
+        }
+        self._perform_export_action_calls_modeladmin_get_queryset_test(data)
+
+    @override_settings(IMPORT_EXPORT_SKIP_ADMIN_EXPORT_UI=True)
+    def test_export_action_calls_modeladmin_get_queryset_skip_export_ui(self):
+        # Issue #1864
+        # Test with specific export items and skip UI
+
+        data = {
+            "format": "0",
+            "export_items": [str(self.cat1.id)],
+            **self.resource_fields_payload,
+        }
+        self._perform_export_action_calls_modeladmin_get_queryset_test(data)
 
     def test_get_export_data_raises_PermissionDenied_when_no_export_permission_assigned(
         self,